Refactor source-decorators -> source-interfaces
