+++ packages/@orbit/data/src/index.ts
@@ -11,8 +11,8 @@ export * from './record';
 export { default as Schema, AttributeDefinition, RelationshipDefinition, KeyDefinition, ModelDefinition, SchemaSettings } from './schema';
 export * from './source';
 export { default as Transform, TransformOrOperations } from './transform';
-export { default as pullable, Pullable, isPullable } from './source-decorators/pullable';
-export { default as pushable, Pushable, isPushable } from './source-decorators/pushable';
-export { default as queryable, Queryable, isQueryable } from './source-decorators/queryable';
-export { default as syncable, Syncable, isSyncable } from './source-decorators/syncable';
-export { default as updatable, Updatable, isUpdatable } from './source-decorators/updatable';
+export { default as pullable, Pullable, isPullable } from './source-interfaces/pullable';
+export { default as pushable, Pushable, isPushable } from './source-interfaces/pushable';
+export { default as queryable, Queryable, isQueryable } from './source-interfaces/queryable';
+export { default as syncable, Syncable, isSyncable } from './source-interfaces/syncable';
+export { default as updatable, Updatable, isUpdatable } from './source-interfaces/updatable';
